+++ HISTORY.rst
@@ -6,3 +6,10 @@ History
 ------------------
 
 * First release on PyPI.
+
+1.1.0 (2018-06-11)
+------------------
+
+* Refactored the API. Now a Single class Conjugator is needed to interface with the module.
+* Includes improved french conjugation model.
+